Move spark actions
